Project Completion - Final Touch Up
Renamed some files as well for consistency.

App is now production ready for demo, and in future I will link real API data into this, just that is full backend task and may have to rebuild components to handle it. So playground right now is where we test models, the rest is demo only and potential features.